Break fetching pollers out to it's own service
